chore: revert refactor seed directory
