+++ apps/api/src/truncateTables.ts
@@ -4,7 +4,7 @@ import postgres from "postgres";
 
 import { seedTruncateAllTables } from "./seed-helpers";
 
-import type { DatabasePg } from "../../common";
+import type { DatabasePg } from "./common";
 
 dotenv.config({ path: "./.env" });
 
@@ -18,6 +18,16 @@ const db = drizzle(sql) as DatabasePg;
 
 async function seed() {
   await seedTruncateAllTables(db);
+  console.log("✨✨✨Created empty database✨✨✨");
+}
+
+if (require.main === module) {
+  seed()
+    .then(() => process.exit(0))
+    .catch((error) => {
+      console.error("An error occurred:", error);
+      process.exit(1);
+    });
 }
 
 export default seed;
